requirements.txt
flask==3.0.0
transformers>=4.50.0
nltk==3.8.1
python-dotenv==1.0.0
azure-ai-textanalytics==5.3.0
torch>=2.2.0
numpy<2.0.0
requests==2.31.0
pyyaml==6.0.1
